elriegocom_app/requirements.txt
alembic==1.14.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
appnope==0.1.4
asttokens==3.0.0
attrs==24.2.0
backports.tarfile==1.2.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudevents==1.11.0
comm==0.2.2
debugpy==1.8.9
decorator==5.1.1
Deprecated==1.2.15
deprecation==2.1.0
distro==1.9.0
docutils==0.21.2
executing==2.1.0
fastapi==0.115.6
Flask==3.1.0
functions-framework==3.8.2
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.24.0
google-api-python-client==2.157.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-appengine-logging==1.5.0
google-cloud-audit-log==0.3.0
google-cloud-core==2.4.1
google-cloud-logging==3.11.3
google-oauth2-tool==0.0.3
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.14.0
grpcio==1.69.0
grpcio-status==1.69.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.30.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.8.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keyring==25.5.0
lazy_loader==0.4
Mako==1.3.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.5.0
narwhals==1.17.0
nest-asyncio==1.6.0
nh3==0.2.19
numpy==2.2.0
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.57.1
opentelemetry-api==1.29.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
pipdeptree==2.16.2
pkginfo==1.10.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.29.1
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.3
pydantic_core==2.27.1
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-multipart==0.0.19
python-socketio==5.11.4
pytz==2024.2
pyzmq==26.2.0
readme_renderer==44.0
redis==5.2.1
referencing==0.35.1
reflex==0.6.6.post3
reflex-chakra==0.6.2
reflex-hosting-cli==0.1.30
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
sqlmodel==0.0.22
stack-data==0.6.3
starlette==0.41.3
starlette-admin==0.14.1
streamlit==1.41.0
tabulate==0.9.0
tenacity==9.0.0
toml==0.10.2
tomlkit==0.13.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
twine==5.1.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.1
watchdog==6.0.0
wcwidth==0.2.13
websockets==14.1
Werkzeug==3.1.3
wrapt==1.17.0
wsproto==1.2.0
zipp==3.21.0